BoxOffice/requirements.txt
chardet==3.0.4
Django==1.11.2
django-braces==1.11.0
django-crispy-forms==1.6.1
django-oauth-toolkit==1.0.0
django-oauth2-provider==0.2.6.1
django-rest-framework-social-oauth2==1.0.8
djangorestframework==3.6.3
httplib2==0.10.3
gunicorn==19.7.1
idna==2.5
IPy==0.83
oauth2client==4.1.1
oauthlib==2.0.2
packaging==16.8
pyasn1==0.2.3
pyasn1-modules==0.0.9
PyJWT==1.5.0
pyparsing==2.2.0
pytz==2017.2
requests==2.18.1
requests-oauthlib==0.8.0
rsa==3.4.2
shortuuid==0.5.0
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
urllib3==1.21.1
whitenoise==3.3.0
django-taggit==0.22.1
django-filter==1.0.2
django-filters==0.1.0
django-taggit-serializer==0.1.5
pandas==0.21.0
psycopg2==2.7.3.2
comtypes==1.1.4
braintree==3.46.0
django-pandas==0.5.1
mysqlclient==1.3.13
google-api-python-client==1.7.4
django-mysql==2.4.1